fix: remove useMutation and useQuery calls from walletProvider
